+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Added: `no-indistinguishable-colors` rule.
 - Added: `at-rule-blockless-newline-after` rule.
 - Fixed: `no-duplicate-selectors` contextualizes selectors by all at-rules, not just media queries.
+- Fixed: `declaration-bang-space-*` understands arbitrary bang declarations (e.g. `!default`).
 - Fixed: `font-weight-notation` now ignore `inherit` value.
 - Fixed: `selector-list-comma-*` rules no longer deliver false positives inside functional notation.